+++ htdocs/includes/stripe/lib/ApiRequestor.php
@@ -25,6 +25,11 @@ class ApiRequestor
     private static $_httpClient;
 
     /**
+     * @var RequestTelemetry
+     */
+    private static $requestTelemetry;
+
+    /**
      * ApiRequestor constructor.
      *
      * @param string|null $apiKey
@@ -37,6 +42,30 @@ class ApiRequestor
             $apiBase = Stripe::$apiBase;
         }
         $this->_apiBase = $apiBase;
+    }
+
+    /**
+     * Creates a telemetry json blob for use in 'X-Stripe-Client-Telemetry' headers
+     * @static
+     *
+     * @param RequestTelemetry $requestTelemetry
+     * @return string
+     */
+    private static function _telemetryJson($requestTelemetry)
+    {
+        $payload = array(
+            'last_request_metrics' => array(
+                'request_id' => $requestTelemetry->requestId,
+                'request_duration_ms' => $requestTelemetry->requestDuration,
+        ));
+
+        $result = json_encode($payload);
+        if ($result != false) {
+            return $result;
+        } else {
+            Stripe::getLogger()->error("Serializing telemetry payload failed!");
+            return "{}";
+        }
     }
 
     /**
@@ -332,6 +361,10 @@ class ApiRequestor
             $defaultHeaders['Stripe-Account'] = Stripe::$accountId;
         }
 
+        if (Stripe::$enableTelemetry && self::$requestTelemetry != null) {
+            $defaultHeaders["X-Stripe-Client-Telemetry"] = self::_telemetryJson(self::$requestTelemetry);
+        }
+
         $hasFile = false;
         $hasCurlFile = class_exists('\CURLFile', false);
         foreach ($params as $k => $v) {
@@ -356,6 +389,8 @@ class ApiRequestor
             $rawHeaders[] = $header . ': ' . $value;
         }
 
+        $requestStartMs = Util\Util::currentTimeMillis();
+
         list($rbody, $rcode, $rheaders) = $this->httpClient()->request(
             $method,
             $absUrl,
@@ -363,6 +398,14 @@ class ApiRequestor
             $params,
             $hasFile
         );
+
+        if (array_key_exists('request-id', $rheaders)) {
+            self::$requestTelemetry = new RequestTelemetry(
+                $rheaders['request-id'],
+                Util\Util::currentTimeMillis() - $requestStartMs
+            );
+        }
+
         return [$rbody, $rcode, $rheaders, $myApiKey];
     }
 
@@ -440,6 +483,16 @@ class ApiRequestor
     public static function setHttpClient($client)
     {
         self::$_httpClient = $client;
+    }
+
+    /**
+     * @static
+     *
+     * Resets any stateful telemetry data
+     */
+    public static function resetTelemetry()
+    {
+        self::$requestTelemetry = null;
     }
 
     /**